Remove the wrappy layer.
