+++ plugin/builtin/hashtransformer/HashTransformer.go
@@ -28,7 +28,7 @@ func (p *plugin) Config(
 func (p *plugin) Transform(m resmap.ResMap) error {
 	for _, res := range m.Resources() {
 		if res.NeedHashSuffix() {
-			h, err := p.hasher.Hash(res)
+			h, err := res.Hash(p.hasher)
 			if err != nil {
 				return err
 			}
